fixed gcc warnings for r4793
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@4794 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinx.cpp
@@ -19003,7 +19003,7 @@ bool CSphIndex_VLN::ParsedMultiQuery ( const CSphQuery * pQuery, CSphQueryResult
 	int64_t iPoolSize = 0;
 	if ( pRanker->ExtraData ( EXTRA_GET_POOL_SIZE, (void**)&iPoolSize ) && iPoolSize>INT_MAX )
 	{
-		pResult->m_sError.SetSprintf ( "ranking factors pool too big (%d Mb), reduce max_matches", iPoolSize/1024/1024 );
+		pResult->m_sError.SetSprintf ( "ranking factors pool too big (%d Mb), reduce max_matches", (int)( iPoolSize/1024/1024 ) );
 		return false;
 	}
